project/requirements.txt
psycopg2
flask
pika
pyjwt
bcrypt